Feat: AI 대화 API 추가

+++ be/src/main/java/com/be/domain/chatbot/request/StartRequest.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 public class StartRequest {
     private boolean isDefault;
-    private Integer defaultId;
+    private Long defaultId;
     private String aiRole;
     private String userRole;
     private String situation;
